+++ src/components/Note.jsx
@@ -1,11 +1,12 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import { useState, useEffect } from "react";
-
+import { useState, useEffect, useCallback } from "react";
+import Swal from "sweetalert2";
+import withReactContent from "sweetalert2-react-content";
 
 function Note(props){
     const navigate = useNavigate();
-
+    const MySwal = withReactContent(Swal)
     const [savedNotes, setSavedNotes] = useState(props.notes);
 
     useEffect(() => {
@@ -13,23 +14,12 @@ function Note(props){
         setSavedNotes(savedArr);
     }, [savedNotes])
 
-    const [favs, setFavs] = useState([]);
-
-    useEffect(() => {
-        const favsInLocal = localStorage.getItem('favs');
-        if(favsInLocal !==null){
-          const favsArray = JSON.parse(favsInLocal)
-          setFavs(favsArray);
-        }
-    }
-    , [setFavs])
-
+    
     let params = new URLSearchParams(document.location.search)
     let title = params.get('title')
     let origin = params.get('from')
     
     let notesArray = props.notes;
-    let favsArray = props.favs;    
 
     let note = props.notes.find(note => note.title === title);
     let noteIdx = props.notes.indexOf(note);
@@ -71,10 +61,44 @@ function Note(props){
         navigate('/', {replace:true})}
     }
     const dispose = () => {
-        notesArray.splice(noteIdx, 1);
-        localStorage.setItem('notes', JSON.stringify(notesArray));
-        navigate('/', {replace:true})
+        MySwal.fire({
+            customClass: {
+                confirmButton: "confirm-btn",
+                popup : "swal-cont"
+            },
+            title : '¿🗑️?',
+            showCancelButton: true,
+            confirmButtonText : <svg xmlns="http://www.w3.org/2000/svg" width="1rem" height="1rem" fill="#fff" viewBox="0 0 24 24"><path d="M9 22l-10-10.598 2.798-2.859 7.149 7.473 13.144-14.016 2.909 2.806z"/></svg> ,
+            cancelButtonText :<svg xmlns="http://www.w3.org/2000/svg" width="1rem" height="1rem" fill="#fff"viewBox="0 0 24 24"><path d="M23 20.168l-8.185-8.187 8.185-8.174-2.832-2.807-8.182 8.179-8.176-8.179-2.81 2.81 8.186 8.196-8.186 8.184 2.81 2.81 8.203-8.192 8.18 8.192z"/></svg> ,
+            denyButtonText: `Don't save`
+        })
+            .then( (result) =>{
+                if(result.isConfirmed){
+                    notesArray.splice(noteIdx, 1);
+                    localStorage.setItem('notes', JSON.stringify(notesArray));
+                    navigate('/', {replace:true})
+                } 
+            })
+        
     }
+
+    const handleKeyPress = useCallback((event) => {
+        if (event.shiftKey === true) {
+          event.key === 's' || event.key === 'S' && save();
+          event.key === 'e' || event.key === 'E' && edit();
+          event.key === 'd' || event.key === 'D' && dispose();
+          event.key === 'Backspace' && navigate('/', {replace:true});
+        }
+
+      }, []);
+
+    useEffect(() => {
+        document.addEventListener('keydown', handleKeyPress);
+        return () => {
+          document.removeEventListener('keydown', handleKeyPress);
+        };
+      }, [handleKeyPress]);
+
     while(note !== null & note !== undefined){
         return(
             <>
